Boingy graph uses hierarchical data type selector
     #3123
