+++ waltz-ng/client/logical-flow/services/logical-flow-utility.js
@@ -124,8 +124,6 @@ export default [
 
 
         return {
-            enrich,
-            getDataTypes,
             buildGraphTweakers,
             enrichDataTypeCounts
         };
