+++ waltz-ng/client/playpen/3/playpen3.html
@@ -1,30 +1,9 @@
 <div>
 
     <waltz-section>
-        <waltz-from-now timestamp="ctrl.time"
-                        days-only="ctrl.daysOnly">
-        </waltz-from-now>
 
-        <br>
-        <br>
-
-        <waltz-from-now timestamp="ctrl.time1"
-                        days-only="ctrl.daysOnly">
-        </waltz-from-now>
-
-        <br>
-        <br>
-
-        <waltz-from-now timestamp="ctrl.time2"
-                        days-only="ctrl.daysOnly">
-        </waltz-from-now>
-
-        <br>
-        <br>
-
-        <waltz-from-now timestamp="ctrl.time3"
-                        days-only="ctrl.daysOnly">
-        </waltz-from-now>
+        <waltz-logical-flows-boingy-graph parent-entity-ref="ctrl.parentEntityRef">
+        </waltz-logical-flows-boingy-graph>
 
 
     </waltz-section>
